+++ tests/test_gates.py
@@ -789,7 +789,7 @@ def test_controlled_globalphase(n_qubits, control_value, tol):
     reason="lightning.kokkos doesn't support zero wires on Windows.",
 )
 @pytest.mark.skipif(
-    device_name in ["lightning.gpu", "lightning.tensor"],
+    device_name in ["lightning.tensor"],
     reason=device_name + " doesn't support zero wires.",
 )
 @pytest.mark.parametrize("control_value", [False, True])
@@ -806,6 +806,65 @@ def test_controlled_globalphase_zero_targetwire(n_qubits, control_value, tol):
             qml.GlobalPhase(0.1234),
             control_wires,
             control_values=([control_value or bool(i % 2) for i, _ in enumerate(control_wires)]),
+        )
+        qml.GlobalPhase(0.512)
+        return qml.state()
+
+    circ = qml.QNode(circuit, dev)
+    circ_def = qml.QNode(circuit, dev_def)
+    assert np.allclose(circ(), circ_def(), tol)
+
+
+@pytest.mark.skipif(
+    (device_name == "lightning.kokkos" and sys.platform == "win32"),
+    reason="lightning.kokkos doesn't support zero wires on Windows.",
+)
+@pytest.mark.skipif(
+    device_name in ["lightning.tensor"],
+    reason=device_name + " cannot be initialized with less than 2 wires.",
+)
+@pytest.mark.parametrize("n_qubits", list(range(1, 3)))
+def test_controlled_globalphase_1ctrl_true_cornercase(n_qubits, tol):
+    """Test that single-controlled (value=True) GlobalPhase with zero-wire is correctly applied to a state"""
+    dev_def = qml.device("default.qubit", wires=n_qubits)
+    dev = qml.device(device_name, wires=n_qubits)
+    control_wires = [0]
+
+    def circuit():
+        qml.StatePrep(get_random_normalized_state(2**n_qubits), wires=range(n_qubits))
+        qml.ctrl(
+            qml.GlobalPhase(0.1234),
+            control_wires,
+            control_values=[True],
+        )
+        return qml.state()
+
+    circ = qml.QNode(circuit, dev)
+    circ_def = qml.QNode(circuit, dev_def)
+    assert np.allclose(circ(), circ_def(), tol)
+
+
+@pytest.mark.skipif(
+    (device_name == "lightning.kokkos" and sys.platform == "win32"),
+    reason="lightning.kokkos doesn't support zero wires on Windows.",
+)
+@pytest.mark.skipif(
+    device_name in ["lightning.tensor"],
+    reason=device_name + " cannot be initialized with less than 2 wires.",
+)
+@pytest.mark.parametrize("n_qubits", list(range(1, 3)))
+def test_controlled_globalphase_1ctrl_false_cornercase(n_qubits, tol):
+    """Test that single-controlled GlobalPhase (value=False) with zero-wire is correctly applied to a state"""
+    dev_def = qml.device("default.qubit", wires=n_qubits)
+    dev = qml.device(device_name, wires=n_qubits)
+    control_wires = [0]
+
+    def circuit():
+        qml.StatePrep(get_random_normalized_state(2**n_qubits), wires=range(n_qubits))
+        qml.ctrl(
+            qml.GlobalPhase(0.1234),
+            control_wires,
+            control_values=[False],
         )
         return qml.state()